dashboard/requirements.txt
python-fasthtml>=0.12.0
httpx>=0.27.0
